refactor: move internal/osm to internal/sources/osm
Pure rename in preparation for the Source abstraction (#60).
No behavior change.
