+++ internal/sources/osm/pbf_test.go
@@ -13,7 +13,7 @@ import (
 )
 
 func TestStreamNodes_FixturePBF(t *testing.T) {
-	f, err := os.Open("../../testdata/andorra-sample.osm.pbf")
+	f, err := os.Open("../../../testdata/andorra-sample.osm.pbf")
 	if err != nil {
 		t.Skipf("fixture PBF not available: %v", err)
 	}
@@ -39,7 +39,7 @@ func TestStreamNodes_FixturePBF(t *testing.T) {
 }
 
 func TestStreamNodes_StopsOnSinkError(t *testing.T) {
-	f, err := os.Open("../../testdata/andorra-sample.osm.pbf")
+	f, err := os.Open("../../../testdata/andorra-sample.osm.pbf")
 	if err != nil {
 		t.Skipf("fixture PBF not available: %v", err)
 	}
